Update the readme

+++ defaults.go
@@ -84,12 +84,12 @@ func (dns *Msg) SetAxfr(z string) {
 // This is only a skeleton Tsig RR that is added as the last RR in the 
 // additional section. The caller should then call TsigGenerate, 
 // to generate the complete TSIG with the secret.
-func (dns *Msg) SetTsig(z, algo string, fudge uint16, timesigned uint64) {
+func (dns *Msg) SetTsig(z, algo string, fudge uint16, timesigned int64) {
 	t := new(RR_TSIG)
 	t.Hdr = RR_Header{z, TypeTSIG, ClassANY, 0, 0}
 	t.Algorithm = algo
 	t.Fudge = 300
-	t.TimeSigned = timesigned
+	t.TimeSigned = uint64(timesigned)
 	dns.Extra = append(dns.Extra, t)
 }
